Moved serialization to separate  files
